add mod export, add SCAD to it
