+++ rewrite/src/lib.rs
@@ -49,3 +49,6 @@ pub mod rules;
 pub mod eval;
 mod permute;
 mod solve;
+
+// Export
+pub mod export;
